requirements.txt
ollama
pandas
pydantic
python-dotenv 